Added swiftformat_update_all macro (#7)
Added swiftformat_update_all macro that defines a target that can be run to format all of the Swift source files and copy them to the workspace directory.
Updated the Github Actions workflow with a scenario that makes a poorly formatted code change, executes the update all target and runs tests. The tests will fail if the update all did not work properly.
Removed the default value for the swift_version attribute. Added it to the swiftformat config file.
Updated the doc for the rules and macros.
Added a build_test so that bazel test //... from the rules_swiftformat workspace works properly.