Added `swiftformat_exclude` to the `rules_swift` helpers and bug fixes. (#19)
Added swiftformat_exclude to swiftformat_library, swiftformat_binary, and swiftformat_test. This allows a Swift file to be included for the build, but not formatted. This is useful for Swift source files that cause SwiftFormat to hang or take a long time to process.
Added missing logic for handling if srcs was not specified for swiftformat_library, swiftformat_binary, or swiftformat_test.
Added exclude_files example which demonstrates the use of the swiftformat_exclude attribute.
Fixed typo in rules_swift_helpers example.
Fixed issue where the name for the formatting targets was not unique.