fix: avoid tests that fail on MacOS runners in CI (#149)
## Background

In February 2024, the following tests started failing in CI on the
GitHub macos-13 runner.

- //examples:exclude_files_test
- //examples:rules_swift_helpers_test

There is no output in the failure. Debugging on the runner shows that
xctest is failing with exit code 83. I have not found a solution to this
problem. We will not run these tests on MacOS.

NOTE: These tests succeed when run locally.

## Fixes

- Run a subset of the integration tests on the MacOS runner.
- Upgrade to Bazel 7.0.2.